[object #177] dissolve 'immutable' module
