--- git-object/src/immutable/mod.rs
@@ -1,2 +1,0 @@
-///
-pub mod object;
