+++ git-object/src/tag/mod.rs
@@ -1,4 +1,4 @@
-use crate::{immutable::object, TagRef};
+use crate::TagRef;
 
 mod decode;
 
@@ -10,10 +10,10 @@ pub mod ref_iter;
 
 impl<'a> TagRef<'a> {
     /// Deserialize a tag from `data`.
-    pub fn from_bytes(data: &'a [u8]) -> Result<TagRef<'a>, object::decode::Error> {
+    pub fn from_bytes(data: &'a [u8]) -> Result<TagRef<'a>, crate::decode::Error> {
         decode::git_tag(data)
             .map(|(_, t)| t)
-            .map_err(object::decode::Error::from)
+            .map_err(crate::decode::Error::from)
     }
     /// The object this tag points to as `Id`.
     pub fn target(&self) -> git_hash::ObjectId {
